mgr/cephadm: move _get_container_image_info to serve.py
